# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      must_change_password:
        type: boolean
      token:
        type: string
      user:
        $ref: '#/definitions/auth.UserResponse'
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  auth.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  health.HealthResponse:
    properties:
      database:
        example: connected
        type: string
      status:
        example: ok
        type: string
      timestamp:
        example: "2025-01-01T00:00:00Z"
        type: string
    type: object
  pagination.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  pagination.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/pagination.Meta'
    type: object
  response.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  trainers.CreateTrainerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      speciality:
        type: string
    required:
    - email
    - name
    - password
    type: object
  trainers.TrainerResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      role:
        type: string
      speciality:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  trainers.UpdateTrainerRequest:
    properties:
      is_active:
        type: boolean
      name:
        type: string
      speciality:
        type: string
    type: object
  users.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@fittrackpro.com
    name: FitTrackPro
  description: API para seguimiento de fitness y entrenamiento personal.
  title: FitTrackPro API
  version: "1.0"
paths:
  /auth/change-password:
    put:
      consumes:
      - application/json
      description: Cambia la contraseña del usuario autenticado. Establece must_change_password
        a false.
      parameters:
      - description: Contraseña actual y nueva
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Cambiar contraseña
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica un usuario y devuelve un token JWT. Si must_change_password
        es true, el usuario debe cambiar su contraseña.
      parameters:
      - description: Credenciales de acceso
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Iniciar sesión
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Crea el primer usuario administrador. Bloquea el registro si ya
        existe un administrador.
      parameters:
      - description: Datos de registro
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Registrar primer administrador
      tags:
      - auth
  /health:
    get:
      description: Verifica que el servidor y la base de datos estén operativos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/health.HealthResponse'
              type: object
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Verificación de estado
      tags:
      - health
  /trainers:
    get:
      description: Lista entrenadores con paginación
      parameters:
      - default: 1
        description: Número de página
        in: query
        name: page
        type: integer
      - default: 20
        description: Registros por página
        in: query
        name: per_page
        type: integer
      - description: Filtrar por estado activo
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pagination.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/trainers.TrainerResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Listar entrenadores
      tags:
      - trainers
    post:
      consumes:
      - application/json
      description: Crea un usuario con rol trainer y su perfil de entrenador
      parameters:
      - description: Datos del entrenador
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/trainers.CreateTrainerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/trainers.TrainerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Crear entrenador (admin)
      tags:
      - trainers
  /trainers/{id}:
    delete:
      description: Realiza borrado lógico del usuario asociado al entrenador
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Eliminar entrenador (admin)
      tags:
      - trainers
    get:
      description: Devuelve un entrenador por su ID
      parameters:
      - description: ID del entrenador
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/trainers.TrainerResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Obtener entrenador
      tags:
      - trainers
    put:
      consumes:
      - application/json
      description: Actualiza los datos del entrenador y del usuario asociado
      parameters:
      - description: ID del entrenador
        in: path
        name: id
        required: true
        type: string
      - description: Datos a actualizar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/trainers.UpdateTrainerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/trainers.TrainerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Actualizar entrenador (admin)
      tags:
      - trainers
  /users:
    get:
      description: Lista usuarios con paginación
      parameters:
      - default: 1
        description: Número de página
        in: query
        name: page
        type: integer
      - default: 20
        description: Registros por página
        in: query
        name: per_page
        type: integer
      - description: Filtrar por estado activo
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pagination.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/users.UserResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Listar usuarios (admin)
      tags:
      - users
  /users/{id}:
    delete:
      description: Realiza borrado lógico del usuario (establece la marca de eliminado)
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Eliminar usuario (admin)
      tags:
      - users
    get:
      description: Devuelve un usuario por su ID
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/users.UserResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Obtener usuario (admin)
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"